do not fail if we receive a datetime.datetime object
If somebody is not using an extract_from option, the upper manager will
try to read and parse the lastrun file, and we are going to get a
datetime.datetime object, that cannot be parsed by dateutil.parser. This
is a quick patch so that we do not fail, but this need to be refactored
though.