Use a list for the tenant mapping
We were using single strings to hold the tenant informatio, so it was
not possible to map several tenants inside a VO. This change uses lists
to that we are able to do such a mapping, using a new element in the
JSON file (i.e. "tenants"). Log a warning about the old format, so that
users can be aware of this.