Fix record extraction and do not request only deleted records
Commit 0058b6cf239f29c8ae7f41212a66dcf9c14f7435 introduced a bug in how
we retrieve the records, since we introduced the "deleted" filter and
this was filtering out all the instances that are not deleted.

Moreover, the way we were retrieving the results was wrong as only
servers with a changed status within the requested period were
considered for the record building.

We cannot use 'changes-since' in the servers.list() API query, as it
will only include changes that have chanted its status after that date.
Also, we cannot just get all the usages and then query the servers.get()
server by server, as deleted servers are not returned by ths
servers.get() call. What we do now is the following.

    1.- List all the deleted servers that changed after the start date
    2.- Build the records for the period [start, end]
    3.- Get all the usages
    4.- Iter over the usages and:
    4.1.- get information for non deleted servers
    4.2.- do nothing with deleted servers, as we collected in in step (2)

However, usage was eventually correctly reported, as sooner or later one
server may have a changed status and we will generate the record.

In order to get accurate results, accounting records must be regenerated
at sites, possibly month by month, like:

    caso-extract  --extract-from 2017-01-01 --extract-to 2017-01-31
    caso-extract  --extract-from 2017-02-01 --extract-to 2017-02-28
    (...)

Closes IFCA/caso#47