Do not lazy-load the extra specs for each flavor
When we access the extra specs for a flavor with the .get_keys() method
the attributes are lazy-loaded each time, meaning that if we have a
large number of servers we are hitting the API for each of those
servers. Since the client does not allow to load the extra specs, we do
it manually, thus hitting the API only when the flavors are loaded for
the first time. By doing so we have passed from 50s to 20s (extracting
700 records).