Do not fail if querying for a server throws an error in the API
Unfortunately, some serves stay in the DB with wrong metadata (probably
old servers, with some information that was not correctly migrated, and
they throw an error when requesting their information. If this happens,
there are no records being generated for the rest of the period, since
we did not catch the exception. This change fixes this.